Add --precompute_aligments_only flag
This flag will cause OpenFold to quit after precomputing alignments.

This can be useful for splitting the CPU-heavy parts of the pipeline
from the GPU-heavy parts. For example, the alignments can be precomputed
on a CPU and IO optimized machine while the actual predictions will be
sent to a GPU cluster.